+++ src/plone/restapi/deserializer/atcontent.py
@@ -24,6 +24,8 @@ class DeserializeFromJson(OrderingMixin, object):
     """JSON deserializer for Archetypes content types
     """
 
+    notifies_create = True
+
     def __init__(self, context, request):
         self.context = context
         self.request = request