+++ src/plone/restapi/services/content/tus.py
@@ -10,7 +10,6 @@ from plone.restapi.exceptions import DeserializationError
 from plone.restapi.interfaces import IDeserializeFromJson
 from plone.restapi.services import Service
 from plone.restapi.services.content.utils import create
-from plone.restapi.services.content.utils import rename
 from plone.restapi.services.content.utils import add
 from plone.rfc822.interfaces import IPrimaryFieldInfo
 from uuid import uuid4
@@ -231,8 +230,6 @@ class UploadPatch(UploadFileBase):
                         filename.lower(), content_type, '') or 'File'
 
                 obj = create(self.context, type_)
-                notify(ObjectCreatedEvent(obj))
-                obj = add(self.context, obj)
             else:
                 obj = self.context
 
@@ -263,7 +260,9 @@ class UploadPatch(UploadFileBase):
                     'Deserialization Error', str(e), 400)
 
             if mode == 'create':
-                rename(obj)
+                if not getattr(deserializer, 'notifies_create', False):
+                    notify(ObjectCreatedEvent(obj))
+                obj = add(self.context, obj)
 
             tus_upload.close()
             tus_upload.cleanup()